interface/ftdi: flash stm32l4x: Update support for the STM32L476
Update the ftdi and mpsse interface to support a wait response.
When in this mode the communication is slowed down, since the
response for each SWD transaction must be looked at.  However,
this seems better than not working at all.

In addition a few changes were made to the STM32L4X flash routines
to correct some corner cases where erasing did not work, as well
as issues with flashing large blocks.

Signed-off-by: David Peyer <David.Peyer@motorola.com>