tcl/board: tcl/interface: tcl/chip: target/swd: Allow reset during SWD connection
Added code to swd_init() to release the reset line if it was enabled
in the configuration files.  In addition if connect_assert_srst is
set in the configuration files assert reset during the connection
procedure.

Added two new Moto HDK MuC configuration files to allow for the MuC
to be reset when the debugger first connects.  This can be used
to wake the MuC from sleep mode when connecting.

Renamed the ap_bridge directory and configuration files to better
match the documentation for the HDK.

Signed-off-by: David Peyer <David.Peyer@motorola.com>