target: Add callback versions of the working area allocate functions
Added versions of the working area allocate functions which
take a callback function and data pointer.  With this change
code downloaded to the target does not have to be cleaned up
immediately after its use.  For flashing this allows multiple
blocks to be programmed without having to download the
programming code each time.  The callback functions can handle
the cleanup when the target is resumed.

New functions were added, as opposed to adding the parameters
to the old functions to allow the existing code using the working
area functions to remain unchanged.

Signed-off-by: David Peyer <David.Peyer@motorola.com>