flash/nor: Add support for generic SPI flash drivers
Added a generic SPI flash driver framework based on mrvlqspi.c with
initial support for the Motorola HSB as well as the STM32 series parts.
The generic driver provides the protocol for SPI flash parts with
each driver providing and init and SPI communications driver.  Testing
was done on the Winbond W25W40EW which was added along with the Winbond
W25W40EW.  The bank sizes of these parts are smaller than most to allow
for more fine control when flashing.

Signed-off-by: David Peyer <David.Peyer@motorola.com>