tcl: board: Add GDB attach handler to halt on connect
Add a GDB attach handler to halt the target on GDB connect.
This is necessary because the APBE's SPI FLASH (bank 1 on MUC
and bank 0 on APBE) needs to access the SPI to read the FLASH's
id. This operation requires the target to be halted.

Also, in the MUC GDB attach handler, call hsb_flash_enable()
to power off the APBE, set-up SPI, and set the SPI switch to
allow the MUC to access the SPI flash.

Initial GDB handler fix from "Jim Wylder <jim@wylder.us>".

Signed-off-by: Mike Corrigan <corrigan@gmail.com>